# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.EmergencyAlarmRequest:
    properties:
      description:
        example: 火灾警报被触发，疑似厨房起火
        type: string
      location:
        example: Building A, Floor 3
        type: string
      property_id:
        description: 物业ID
        example: 1
        type: integer
      reported_by:
        description: 报告人ID
        example: 1
        type: integer
      type:
        description: 如：fire(火灾)、intrusion(入侵)、medical(医疗)等
        example: fire
        type: string
    required:
    - location
    - property_id
    - type
    type: object
  controllers.EmergencyNotificationRequest:
    properties:
      content:
        example: A栋3楼发生火灾，请所有居民立即疏散。
        type: string
      expires_at:
        description: 可选，不提供则默认24小时
        example: "2023-07-01T15:00:00Z"
        type: string
      is_public:
        description: 是否为公开通知
        example: false
        type: boolean
      property_id:
        description: 关联的物业ID，可以为空表示全局通知
        example: 1
        type: integer
      severity:
        description: high, medium, low
        example: high
        type: string
      target_type:
        description: all, residents, staff
        example: all
        type: string
      title:
        example: 紧急通知：小区火灾警报
        type: string
    required:
    - content
    - severity
    - title
    type: object
  controllers.EmergencyUnlockRequest:
    properties:
      reason:
        example: 火灾疏散
        type: string
    required:
    - reason
    type: object
...
  controllers.ErrorResponse:
    properties:
      code:
        example: 401
        type: integer
      data: {}
      message:
        example: Invalid username or password
        type: string
    type: object
...
host: 39.108.49.167:20033
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: A comprehensive door access management system with video calling capabilities
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: intercom_http_service API
  version: "1.0"
paths:
...
  /emergency/alarm:
    post:
      consumes:
      - application/json
      description: Trigger an emergency alarm and notify relevant personnel
      parameters:
      - description: Alarm request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.EmergencyAlarmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Trigger Emergency Alarm
      tags:
      - Emergency
  /emergency/contacts:
    get:
      consumes:
      - application/json
      description: Get a list of all emergency contacts in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Emergency Contacts
      tags:
      - Emergency
  /emergency/notify-all:
    post:
      consumes:
      - application/json
      description: Send notification to all users during an emergency
      parameters:
      - description: Notification request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.EmergencyNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Notify All Users
      tags:
      - Emergency
  /emergency/unlock-all:
    post:
      consumes:
      - application/json
      description: Unlock all doors in the system during an emergency
      parameters:
      - description: Unlock request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.EmergencyUnlockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Emergency Unlock All Doors
      tags:
      - Emergency
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"